refactor(tokens-parser): captureAnalyzedTokensReferenceErrors utils
